python-impl/requirements.txt
langgraph>=0.3.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langsmith>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
redis>=5.2.0
faiss-cpu>=1.9.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
mcp>=1.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
numpy>=1.26.0
